Call the single, not double-precision version of copysign for the float arguments.

+++ lib/libm/src/s_cprojf.c
@@ -1,4 +1,4 @@
-/*	$OpenBSD: s_cprojf.c,v 1.2 2010/07/19 00:04:07 guenther Exp $	*/
+/*	$OpenBSD: s_cprojf.c,v 1.3 2011/05/31 21:13:44 martynas Exp $	*/
 /*
  * Copyright (c) 2008 Martynas Venckus <martynas@openbsd.org>
  *
@@ -25,7 +25,7 @@ cprojf(float complex z)
 
 	if (isinf(__real__ z) || isinf(__imag__ z)) {
 		__real__ res = INFINITY;
-		__imag__ res = copysign(0.0, __imag__ z);
+		__imag__ res = copysignf(0.0, __imag__ z);
 	} else {
 		res = z;
 	}
